Remove old tools, update readme for specific Docker version (#6)
We may or may not need a specific version of Docker. The current setup installs a cask of docker which is usually the most up to date version so we can't rely on that. This PR add's a step to manually install docker and also removes old unused tools that are causing issues.